refactor: move src/http/__tests__ to src/adapters/http/__tests__
