+++ src/adapters/http/__tests__/insights.e2e.pg.test.ts
@@ -1,8 +1,8 @@
 import { afterAll, afterEach, describe, expect, it } from "vitest";
-import { buildApp } from "../../app.js";
-import type { Db } from "../../ledger/pg/db.js";
-import { createDb } from "../../ledger/pg/db.js";
-import { clmmInsights } from "../../ledger/pg/schema/index.js";
+import { buildApp } from "../../../app.js";
+import type { Db } from "../../../ledger/pg/db.js";
+import { createDb } from "../../../ledger/pg/db.js";
+import { clmmInsights } from "../../../ledger/pg/schema/index.js";
 
 const PG_CONNECTION_STRING =
   process.env.DATABASE_URL ?? "postgres://test:test@localhost:5432/regime_engine_test";
