+++ src/adapters/http/__tests__/regimeCurrent.e2e.test.ts
@@ -2,9 +2,9 @@ import { rmSync } from "node:fs";
 import { join } from "node:path";
 import { tmpdir } from "node:os";
 import { afterEach, describe, expect, it } from "vitest";
-import { buildApp } from "../../app.js";
-import { createLedgerStore, getLedgerCounts } from "../../ledger/store.js";
-import { MARKET_REGIME_CONFIG } from "../../engine/marketRegime/config.js";
+import { buildApp } from "../../../app.js";
+import { createLedgerStore, getLedgerCounts } from "../../../ledger/store.js";
+import { MARKET_REGIME_CONFIG } from "../../../engine/marketRegime/config.js";
 
 const FIFTEEN_MIN_MS = 15 * 60 * 1000;
 const ONE_HOUR_MS = 60 * 60 * 1000;
