+++ src/adapters/http/__tests__/srLevels.e2e.test.ts
@@ -2,8 +2,8 @@ import { rmSync } from "node:fs";
 import { join } from "node:path";
 import { tmpdir } from "node:os";
 import { afterEach, describe, expect, it } from "vitest";
-import { buildApp } from "../../app.js";
-import { createLedgerStore, getLedgerCounts } from "../../ledger/store.js";
+import { buildApp } from "../../../app.js";
+import { createLedgerStore, getLedgerCounts } from "../../../ledger/store.js";
 
 const createdDbPaths: string[] = [];
